requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Pillow==10.1.0
rembg==2.0.50
git+https://github.com/openai/CLIP.git
torch==2.1.0
torchvision==0.16.0
opencv-python-headless==4.8.1.78
scikit-learn==1.3.2
numpy==1.26.2
gunicorn==21.2.0